feat: Enhance StockChat with comprehensive Indian stock market support
- Add intelligent Indian stock symbol and company name detection
- Implement development chat endpoint with real-time stock data via Yahoo Finance
- Support both NSE symbols (.NS) and ADR mappings with clear exchange labeling
- Enhanced symbol detection recognizes company names (e.g., "HDFC Bank" → HDB)
- Add fallback logic: NSE → ADR → BSE for maximum data availability
- Clear user context: distinguish between Indian market data vs ADR prices
- Comprehensive mapping for 50+ major Indian companies across sectors
- Smart filtering to avoid false positives from common English words
- Fix MongoDB data types from Decimal to float for better compatibility

Key Features:
✅ Natural language queries: "HDFC Bank price", "What's Infosys worth?"
✅ Company name recognition: "Tata Consultancy Services" → TCS
✅ Exchange transparency: Clear ADR vs NSE/BSE labeling
✅ Real-time data: Current prices, changes, market cap from Yahoo Finance
✅ Fallback support: Tries multiple exchanges for data availability

+++ app/models/stock.py
@@ -10,19 +10,19 @@ class Stock(Document):
     exchange: str = Field(default="NSE")
     sector: Optional[str] = None
     industry: Optional[str] = None
-    market_cap: Optional[Decimal] = None
-    current_price: Optional[Decimal] = None
-    price_change: Optional[Decimal] = None
-    price_change_percent: Optional[Decimal] = None
+    market_cap: Optional[float] = None
+    current_price: Optional[float] = None
+    price_change: Optional[float] = None
+    price_change_percent: Optional[float] = None
     last_updated: datetime = Field(default_factory=datetime.utcnow)
     is_active: bool = Field(default=True)
     
     # Financial metrics
-    pe_ratio: Optional[Decimal] = None
-    pb_ratio: Optional[Decimal] = None
-    dividend_yield: Optional[Decimal] = None
-    roe: Optional[Decimal] = None
-    debt_to_equity: Optional[Decimal] = None
+    pe_ratio: Optional[float] = None
+    pb_ratio: Optional[float] = None
+    dividend_yield: Optional[float] = None
+    roe: Optional[float] = None
+    debt_to_equity: Optional[float] = None
     
     class Settings:
         collection = "stocks"
